requirements.txt
numpy==1.21.6
matplotlib==3.5.2
Cython==0.29.30
opencv-python==4.5.5.64
Pillow==9.1.1
setuptools==62.3.1